new routes + new tests
